backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
torch>=2.0.0
transformers>=4.36.0
pillow>=10.0.0
numpy>=1.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=1.8.2 